+++ src/lib/hashset/hset.cpp
@@ -47,7 +47,7 @@ const void* sfhash_hashset_sha2_256(const SFHASH_Hashset* hset) {
   return hset->holder.fhdr.sha2_256.data();
 }
 
-size_t sfhash_hashset_count(
+uint64_t sfhash_hashset_count(
   const SFHASH_Hashset* hset,
   size_t tidx)
 {
